pdf-extract/requirements.txt
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
pdfminer.six==20221105
pdfplumber==0.10.4
pillow==10.2.0
pycparser==2.21
PyPDF2==3.0.1
pypdfium2==4.27.0
python-dotenv==1.0.1
